Cache 32 to 64 bit (#62)
* switched from saving/reading 32 bit unsigned integers to the cache to using 64 bit ones

* added debug code to verify that we're loading all the visits

+++ src/Locations.C
@@ -94,9 +94,9 @@ void Locations::loadLocationData() {
   }
   
   // Find starting line for our data through location cache.
-  locationCache.seekg(thisIndex * sizeof(uint32_t));
-  uint32_t locationOffset;
-  locationCache.read((char *) &locationOffset, sizeof(uint32_t));
+  locationCache.seekg(thisIndex * sizeof(uint64_t));
+  uint64_t locationOffset;
+  locationCache.read((char *) &locationOffset, sizeof(uint64_t));
   locationData.seekg(locationOffset);
 
   // Read in our location data.
